refactor(shapescript): move the render TOOL into the plugin as well
1.2.0 shared the renderer and left the tool wrapper behind. Porting
`renderShapeScript` to MulmoTerminal made that immediately visible: what the
second host would have had to copy is the schema, the defaults, the four-view
sheet and the sentence naming the saved file — which is to say, every part a
MODEL sees. Those are the worst things to keep two copies of, because drift in
them is invisible until an agent behaves differently in the two apps.

So `executeRenderShapeScript(deps, args)` joins the renderer in
`@mulmoclaude/shapescript-plugin/render`, and a host supplies only what is
genuinely its own:

  readShape  — read a `.shape` through THIS host's file layer
  saveImage  — persist the PNG and answer with the path the AGENT should read
  onWarning  — optional logger

`saveImage` returning the path rather than taking one is deliberate: MulmoClaude
saves into `artifacts/images/` and can answer relatively, while MulmoTerminal
runs sessions in per-project directories where only an absolute path resolves.
Only the host knows which, so only the host decides.

This host's tool file drops from 180 lines to a definition plus two callbacks.
Behaviour is unchanged — same schema, same budgets, same messages.
