+++ server/agent/mcp-tools/renderShapeScript.ts
@@ -5,68 +5,30 @@
 // pushed to the chat canvas, and the answer is a file path — the same contract
 // `generateImage` uses, which is what makes the result readable (the image
 // lands under `artifacts/images/` and the agent opens it with Read).
+//
+// The tool itself — schema, defaults, the four-view sheet, the sentence naming
+// the saved file — lives in `@mulmoclaude/shapescript-plugin/render`, because
+// MulmoTerminal offers the same tool and those are precisely the parts that
+// would drift if each host kept a copy. What is left here is what is actually
+// this host's: reading a `.shape` through its file layer, saving into its image
+// store, and its logger.
 
-import { isShapeArtifactPath, isPresentableShapePath, toArtifactsRelative } from "@mulmoclaude/shapescript-plugin";
+import {
+  executeRenderShapeScript,
+  RENDER_SHAPE_SCRIPT_DESCRIPTION,
+  RENDER_SHAPE_SCRIPT_PROMPT,
+  RENDER_SHAPE_SCRIPT_SCHEMA,
+  RENDER_SHAPE_SCRIPT_TOOL_NAME,
+  RENDER_TOOL_TIMEOUT_MS,
+} from "@mulmoclaude/shapescript-plugin/render";
+import { isShapeArtifactPath, isPresentableShapePath, toArtifactsRelative, SHAPE_EXTENSIONS } from "@mulmoclaude/shapescript-plugin";
 import { readFile } from "fs/promises";
 import path from "node:path";
 import { saveImage } from "../../utils/files/image-store.js";
 import { resolveByPath } from "../../utils/files/by-path.js";
 import { workspacePath } from "../../workspace/workspace.js";
-import { renderShapeScriptSheet, RenderUnavailableError, RENDER_BUDGET_MS } from "@mulmoclaude/shapescript-plugin/render";
-import type { ViewAngle } from "@mulmoclaude/shapescript-plugin/render";
-import { errorMessage } from "../../utils/errors.js";
 import { log } from "../../system/logger/index.js";
-import { ONE_SECOND_MS } from "../../utils/time.js";
 import type { McpTool } from "./index.js";
-
-const DEFAULT_AZIMUTH = 30;
-const DEFAULT_ELEVATION = 25;
-const DEFAULT_ZOOM = 1;
-const DEFAULT_TILE = 480;
-// A tile bigger than this buys the model nothing and costs render time; the
-// sheet is up to 2x2 of these.
-const MAX_TILE = 900;
-const MIN_TILE = 160;
-// Not a true 90: a hair off vertical keeps a little of the sides visible, which
-// is what tells the reader how tall the model is.
-const TOP_ELEVATION = 85;
-
-const SHAPE_EXTENSIONS = [".shape"] as const;
-
-function clampTile(value: unknown, fallback: number): number {
-  if (typeof value !== "number" || !Number.isFinite(value)) return fallback;
-  return Math.min(MAX_TILE, Math.max(MIN_TILE, Math.round(value)));
-}
-
-function num(value: unknown, fallback: number): number {
-  return typeof value === "number" && Number.isFinite(value) ? value : fallback;
-}
-
-/** Round for a caption — an azimuth of `30.000000000000004` helps nobody. */
-const deg = (value: number): string => `${Math.round(value)}°`;
-
-/**
- * The angles one call renders.
- *
- * The default is FOUR views, not one, and that is the point of the tool: a
- * single projection is ambiguous about depth and occlusion, and a model asked
- * to judge one confidently gets "which of these is in front" wrong. Three
- * quarters around the model plus a top-down resolves it.
- */
-export function viewAngles(azimuth: number, elevation: number, single: boolean): ViewAngle[] {
-  const view = (turn: number, rise: number, name: string): ViewAngle => ({
-    azimuth: turn,
-    elevation: rise,
-    label: `${name} — az ${deg(turn)}, el ${deg(rise)}`,
-  });
-  if (single) return [view(azimuth, elevation, "view")];
-  return [
-    view(azimuth, elevation, "front-right"),
-    view(azimuth + 90, elevation, "back-right"),
-    view(azimuth + 180, elevation, "back-left"),
-    view(azimuth, TOP_ELEVATION, "top"),
-  ];
-}
 
 /** Read the source a `path` argument names — an `artifacts/shapes/**` file this
  *  tool saved, or any other `.shape` on disk. Mirrors the plugin's own routing
@@ -81,101 +43,28 @@ async function readShapeFile(filePath: string): Promise<string> {
   return readFile(resolved, "utf-8");
 }
 
-async function resolveScript(args: Record<string, unknown>): Promise<string> {
-  const script = typeof args.script === "string" && args.script.trim() !== "" ? args.script : undefined;
-  const filePath = typeof args.path === "string" && args.path.trim() !== "" ? args.path : undefined;
-  if (script && filePath) throw new Error("Provide either `script` or `path`, not both");
-  if (script) return script;
-  if (filePath) return readShapeFile(filePath);
-  throw new Error("Provide either `script` (inline source) or `path` (an existing .shape file)");
-}
-
 export const renderShapeScript: McpTool = {
   definition: {
-    name: "renderShapeScript",
-    description:
-      "Render a ShapeScript model to a PNG image and save it, so you can LOOK at the model you wrote and check it before showing it to the user. Returns the image path — read that file to see the result. By default it renders four camera angles onto one sheet, because a single view cannot settle what is in front of what. Takes the same source as presentShapeScript: inline `script`, or `path` to a saved .shape file.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        script: {
-          type: "string",
-          description: "ShapeScript source to render. Provide either this or `path`, not both.",
-        },
-        path: {
-          type: "string",
-          description: "Path to an existing .shape file to render (e.g. one presentShapeScript saved under artifacts/shapes/).",
-        },
-        azimuth: {
-          type: "number",
-          description: `Camera rotation around the model in degrees, 0 = looking from +Z. Default ${DEFAULT_AZIMUTH}. In the default four-view sheet this is the FIRST view's angle; the others are offset from it.`,
-        },
-        elevation: {
-          type: "number",
-          description: `Camera height in degrees above the horizon, 90 = straight down. Default ${DEFAULT_ELEVATION}. Avoid 0 and 90 for the main view — an axis-aligned shot flattens depth.`,
-        },
-        zoom: {
-          type: "number",
-          description: `Multiplier on the automatic framing, which fits the whole model. 1 = fit (default), 2 = twice as close, 0.5 = further out. Prefer this over guessing a camera distance: it means the same thing at any model scale.`,
-        },
-        views: {
-          type: "string",
-          enum: ["quad", "single"],
-          description:
-            "`quad` (default) renders four angles on one sheet — use it to judge shape and layout. `single` renders only `azimuth`/`elevation`, for a close look at one detail.",
-        },
-        projection: {
-          type: "string",
-          enum: ["perspective", "orthographic"],
-          description: "`perspective` (default) looks natural; `orthographic` keeps parallel edges parallel, which is better for comparing proportions.",
-        },
-        width: {
-          type: "number",
-          description: `Pixel width of ONE view. Default ${DEFAULT_TILE}, clamped to ${MIN_TILE}–${MAX_TILE}.`,
-        },
-        height: {
-          type: "number",
-          description: `Pixel height of ONE view. Default ${DEFAULT_TILE}, clamped to ${MIN_TILE}–${MAX_TILE}.`,
-        },
-      },
-      required: [],
-    },
+    name: RENDER_SHAPE_SCRIPT_TOOL_NAME,
+    description: RENDER_SHAPE_SCRIPT_DESCRIPTION,
+    inputSchema: RENDER_SHAPE_SCRIPT_SCHEMA,
   },
   // Launching a browser and rasterising several views outlasts the bridge's
   // 30 s default for external-API tools; without the headroom a render that was
-  // about to succeed is aborted in transit. `RENDER_BUDGET_MS` already covers
-  // launch + render, so this margin is for the work either side of it —
-  // serialising the scene and writing the PNG.
-  bridgeTimeoutMs: RENDER_BUDGET_MS + 30 * ONE_SECOND_MS,
-  prompt:
-    "Use renderShapeScript to CHECK a 3D model you wrote before presenting it — it saves a PNG and returns the path, which you then read to see what the model actually looks like. Fix what you see and re-render. Rendering needs a local headless browser; if it reports one is missing, say so and continue with presentShapeScript rather than retrying.",
+  // about to succeed is aborted in transit.
+  bridgeTimeoutMs: RENDER_TOOL_TIMEOUT_MS,
+  prompt: RENDER_SHAPE_SCRIPT_PROMPT,
   handler: async (args: Record<string, unknown>): Promise<string> => {
-    const script = await resolveScript(args);
-    const azimuth = num(args.azimuth, DEFAULT_AZIMUTH);
-    const elevation = num(args.elevation, DEFAULT_ELEVATION);
-    const views = viewAngles(azimuth, elevation, args.views === "single");
-    log.info("render", "renderShapeScript: start", { views: views.length, bytes: script.length });
-    try {
-      const base64 = await renderShapeScriptSheet({
-        script,
-        onWarning: (message) => log.warn("render", message),
-        views,
-        width: clampTile(args.width, DEFAULT_TILE),
-        height: clampTile(args.height, DEFAULT_TILE),
-        zoom: Math.max(num(args.zoom, DEFAULT_ZOOM), 0.01),
-        projection: args.projection === "orthographic" ? "orthographic" : "perspective",
-      });
-      const imagePath = await saveImage(base64);
-      log.info("render", "renderShapeScript: ok", { imagePath });
-      const captions = views.map((angle) => angle.label).join("; ");
-      const count = views.length === 1 ? "1 view" : `${views.length} views`;
-      return `Saved render to ${imagePath} (${count}: ${captions}). Read that file to see the model.`;
-    } catch (err) {
-      if (err instanceof RenderUnavailableError) {
-        log.warn("render", "renderShapeScript: unavailable", { error: errorMessage(err) });
-        return err.message;
-      }
-      throw err;
-    }
+    log.info("render", "renderShapeScript: start", { args: Object.keys(args).join(",") });
+    const message = await executeRenderShapeScript(
+      {
+        readShape: readShapeFile,
+        saveImage,
+        onWarning: (warning) => log.warn("render", warning),
+      },
+      args,
+    );
+    log.info("render", "renderShapeScript: done");
+    return message;
   },
 };
