+++ test/agent/test_renderShapeScript.ts
@@ -6,8 +6,8 @@
 import { describe, it } from "node:test";
 import assert from "node:assert/strict";
 
-import { renderShapeScript, viewAngles } from "../../server/agent/mcp-tools/renderShapeScript.js";
-import { gridFor } from "@mulmoclaude/shapescript-plugin/render";
+import { renderShapeScript } from "../../server/agent/mcp-tools/renderShapeScript.js";
+import { gridFor, viewAngles } from "@mulmoclaude/shapescript-plugin/render";
 
 describe("renderShapeScript view angles", () => {
   // The default is four views, and that IS the feature: one projection is
